Avoid exception when deleting lease
After a lease is completed, the on_end action is run. But it is also
called again when deleting the lease, which throws an exception because
it can't find the nova aggregate which has already been deleted.

The strack trace looks like this:

[-] Exception during message handling: Aggregate '9' not found!
Traceback (most recent call last):
  File "/usr/lib/python2.7/site-packages/oslo/messaging/rpc/dispatcher.py", line 134, in _dispatch_and_reply
    incoming.message))
  File "/usr/lib/python2.7/site-packages/oslo/messaging/rpc/dispatcher.py", line 177, in _dispatch
    return self._do_dispatch(endpoint, method, ctxt, args)
  File "/usr/lib/python2.7/site-packages/oslo/messaging/rpc/dispatcher.py", line 123, in _do_dispatch
    result = getattr(endpoint, method)(ctxt, **new_args)
  File "/usr/lib/python2.7/site-packages/climate/utils/service.py", line 76, in run_method
    return method(**kwargs)
  File "/usr/lib/python2.7/site-packages/climate/manager/service.py", line 397, in delete_lease
    plugin.on_end(reservation['resource_id'])
  File "/usr/lib/python2.7/site-packages/climate/plugins/oshosts/host_plugin.py", line 197, in on_end
    pool.delete(reservation['resource_id'])
  File "/usr/lib/python2.7/site-packages/climate/plugins/oshosts/reservation_pool.py", line 135, in delete
    agg = self.get_aggregate_from_name_or_id(pool)
  File "/usr/lib/python2.7/site-packages/climate/plugins/oshosts/reservation_pool.py", line 84, in get_aggregate_from_name_or_id
    raise manager_exceptions.AggregateNotFound(pool=aggregate_obj)
AggregateNotFound: Aggregate '9' not found!

 [-] Returning exception Aggregate '9' not found! to caller

+++ climate/plugins/oshosts/host_plugin.py
@@ -188,27 +188,29 @@ class PhysicalHostPlugin(base.BasePlugin, nova.NovaClientWrapper):
         reservations = db_api.reservation_get_all_by_values(
             resource_id=resource_id)
         for reservation in reservations:
+            if reservation['status'] not in ['completed', 'deleted']:
+                allocations = db_api.host_allocation_get_all_by_values(
+                    reservation_id=reservation['id'])
+                pool = rp.ReservationPool()
+                for allocation in allocations:
+                    db_api.host_allocation_destroy(allocation['id'])
+                    hyp = self.nova.hypervisors.get(
+                            self._get_hypervisor_from_name_or_id(
+                            allocation['compute_host_id'])
+                    )
+                    if hyp.__dict__['running_vms'] > 0:
+                        hyp = self.nova.hypervisors.search(hyp.__dict__['hypervisor_hostname'], servers=True)
+                        for server in hyp[0].__dict__['servers']:
+                            s = self.nova.servers.get(server['uuid'])
+                            s.delete()
+                pool.delete(reservation['resource_id'])
+
             db_api.reservation_update(reservation['id'],
                                       {'status': 'completed'})
             host_reservation = db_api.host_reservation_get_by_reservation_id(
                 reservation['id'])
             db_api.host_reservation_update(host_reservation['id'],
                                            {'status': 'completed'})
-            allocations = db_api.host_allocation_get_all_by_values(
-                reservation_id=reservation['id'])
-            pool = rp.ReservationPool()
-            for allocation in allocations:
-                db_api.host_allocation_destroy(allocation['id'])
-                hyp = self.nova.hypervisors.get(
-                        self._get_hypervisor_from_name_or_id(
-                        allocation['compute_host_id'])
-                )
-                if hyp.__dict__['running_vms'] > 0:
-                    hyp = self.nova.hypervisors.search(hyp.__dict__['hypervisor_hostname'], servers=True)
-                    for server in hyp[0].__dict__['servers']:
-                        s = self.nova.servers.get(server['uuid'])
-                        s.delete()
-            pool.delete(reservation['resource_id'])
 
     def _get_extra_capabilities(self, host_id):
         extra_capabilities = {}
